Move test object files into assets folder
