+++ react-app/src/assets/GAD7.js
