+++ react-app/src/assets/index.js
